add unified graph source display module; redact secrets everywhere
