+++ fluree-db-cli/src/commands/list.rs
@@ -1,6 +1,7 @@
 use crate::config::{self, TomlSyncConfigStore};
 use crate::context;
 use crate::error::{CliError, CliResult};
+use crate::graph_source_display::format_source_type;
 use colored::Colorize;
 use comfy_table::{ContentArrangement, Table};
 use fluree_db_api::server_defaults::FlureeDir;
@@ -220,15 +221,4 @@ fn print_ledger_list(result: &serde_json::Value, remote_label: Option<&str>) -> 
 
     println!("{table}");
     Ok(())
-}
-
-fn format_source_type(st: &fluree_db_nameservice::GraphSourceType) -> String {
-    match st {
-        fluree_db_nameservice::GraphSourceType::Bm25 => "BM25".to_string(),
-        fluree_db_nameservice::GraphSourceType::Vector => "Vector".to_string(),
-        fluree_db_nameservice::GraphSourceType::Geo => "Geo".to_string(),
-        fluree_db_nameservice::GraphSourceType::R2rml => "R2RML".to_string(),
-        fluree_db_nameservice::GraphSourceType::Iceberg => "Iceberg".to_string(),
-        fluree_db_nameservice::GraphSourceType::Unknown(s) => format!("Unknown({s})"),
-    }
 }
